Finish The Game Development
 - Make Weapon Better :
   SniperRifle(Stronger)、ShootGun(Multiple)、CustomWeapon(FireBallGun)
 - Custom Enemies & Enemy Edition(AI Editing)
 - Custom Weapon & Weapon Edition
 - Scene Edit (Super Difficult)

+++ Assets/FPS/Scripts/Game/Shared/Health.cs
@@ -77,7 +77,7 @@ namespace Unity.FPS.Game
                 return;
 
             // call OnDie action
-            if (CurrentHealth <= 0f)
+            if ((int)(CurrentHealth) <= 0)
             {
                 m_IsDead = true;
                 OnDie?.Invoke();
